Update README.md
Here's a note regarding Issue #43 - I tried to leave it generic, but with a note regarding the async module specifically.  

Once async module is resolved, it may not make sense to have this addition; but I left it so that the async sentence can deleted if the general cache tip is still relevant.

Of course, definitely feel free to reject this pull request - it may inspire a better way of incorporating into your documentation.